Improved loading process
This allows using the lib within a bundler, while still providing an
autoloading mechanism for standard usage.
Translations are migrated from yaml to ts to simplify the build and
import process, as there is no real advantage to using yaml here.
The preloading mechanism is removed, as there is a 99% chance that the
UI will be shown on a first visit.