Simplified banner styles
The default banner styles don't depend on the screen width anymore.
This avoids a potential mismatch between the media queries and the
public configuration variables, causing the banner to overflow in some
cases.
The distinction between regular and "forced" banner is also removed, as
the backdrop on forced banners is distictive enough.