Distributing standalone modules
This commit is obviously too big but also hard to split.
The main goal is to build standalone modules instead of using a
complicated async loading workflow, as to ease development,
optimization, and usage.
This also include a huge rewrite of the docs to accommodate for the
change and make them more readable.